requirements.txt
amqp==5.0.6
arrow==1.2.1
asgiref==3.4.1
billiard==3.6.4.0
blessed==1.19.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
dataclasses==0.6
defusedxml==0.7.1
Django==3.2.6
django-admin==2.0.1
django-allauth==0.45.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-crontab==0.7.1
django-dbbackup==3.3.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-picklefield==3.0.1
django-redis==5.0.0
django-six==1.0.4
django-timezone-field==4.2.1
djangorestframework==3.12.4
Flask==2.0.1
flower==1.0.0
future==0.18.2
humanize==3.12.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
Pillow==8.3.2
prometheus-client==0.12.0
prompt-toolkit==3.0.22
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.1
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
xlwt==1.3.0
zipp==3.5.0
